+++ GUI-dev/src/app/pages/pages-menu.ts
@@ -12,6 +12,11 @@ export const MENU_ITEMS: NbMenuItem[] = [
     icon: 'ion-cash',
     link: '/processstore-search',
   },
+  {
+    title: 'Approval Page',
+    icon: 'nb-checkmark',
+    link: '/approval',
+  },
   /*{
     title: 'FEATURES',
     group: true,