+++ GUI-dev/src/app/pages/pages.module.ts
@@ -5,6 +5,8 @@ import { PagesRoutingModule } from './pages-routing.module';
 import { ThemeModule } from '../@theme/theme.module';
 import {HomeModule} from './home/home.module';
 import { ProcessStoreSearchModule } from './processstore-search/processstore-search.module';
+import {ApprovalModule} from './approval/approval.module';
+import {ApprovalDetailsModule} from './approval-details/approval-details.module';
 
 import AdminModule from './admin/admin.module';
 
@@ -20,6 +22,8 @@ const PAGES_COMPONENTS = [
     ThemeModule,
     DashboardModule,
     HomeModule,
+    ApprovalModule,
+    ApprovalDetailsModule,
     AdminModule,
     UserDetailsModule,