+++ GUI-dev/src/models/models.ts
@@ -3,6 +3,15 @@
 
 export class Process {
   // define props...
+  process_id?: number;
+  created_at?: Date;
+  price?: number;
+  process_description?: string;
+  process_name?: string;
+  state?: string;
+  version?: number;
+  u_id?: number;
+  creator?: string;
 }
 
 export class StoreProcess {
@@ -16,6 +25,21 @@ export class StoreProcess {
   // missing in current ProcessStoreDTO in backend:
   processAverageRating: number;
 }
+
+
+
+export class Review {
+  // define props...
+  review_id?: number;
+  comment?: string;
+  is_approved?: number;
+  u_id?: number;
+  process_id: number;
+  approver: number;
+  uploader?: string;
+  created_at?: Date;
+}
+
 
 
 export class User {